# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads/{id}/messages:
    get:
      operationId: getLeadMessages
      tags: [messaging]
      summary: Get outbound message history for a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of outbound messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OutboundMessage"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /leads/{id}/messages/draft:
    post:
      operationId: draftLeadMessage
      tags: [messaging]
      summary: AI-draft a message for a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DraftMessageInput"
      responses:
        "200":
          description: Drafted message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageDraft"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /leads/{id}/messages/send:
    post:
      operationId: sendLeadMessage
      tags: [messaging]
      summary: Send a message to a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageInput"
      responses:
        "201":
          description: Sent message record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutboundMessage"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    ClerkAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk session JWT. The browser sends this as a cookie managed by the Clerk SDK; mobile clients send it in the Authorization header as `Bearer <token>`.
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    OutboundMessage:
      type: object
      properties:
        id:
          type: integer
        leadId:
          type: integer
        channel:
          type: string
          enum: [email, sms]
        subject:
          type: string
          nullable: true
        body:
          type: string
        status:
          type: string
          enum: [sent, failed, pending]
        errorMessage:
          type: string
          nullable: true
        sentAt:
          type: string
          nullable: true
        sequenceStep:
          type: integer
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - leadId
        - channel
        - body
        - status
        - createdAt
    DraftMessageInput:
      type: object
      properties:
        channel:
          type: string
          enum: [email, sms]
        stepHint:
          type: integer
          nullable: true
      required:
        - channel
    MessageDraft:
      type: object
      properties:
        subject:
          type: string
          nullable: true
        body:
          type: string
        channel:
          type: string
          enum: [email, sms]
      required:
        - body
        - channel
    SendMessageInput:
      type: object
      properties:
        channel:
          type: string
          enum: [email, sms]
        subject:
          type: string
          nullable: true
        body:
          type: string
        sequenceStep:
          type: integer
          nullable: true
      required:
        - channel
        - body
...